requirements.txt
accelerate==0.26.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
et-xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.4
filelock==3.16.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.106
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.10
packaging==24.1
pdfminer-six==20240706
pillow==11.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.2
sympy==1.13.1
tenacity==8.1.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
